test: 고수가 가진 서비스 조회 SubItem->Expert 도메인으로 이동 후 테스트 코드 작성(GS-70)
