+++ src/test/java/com/foo/gosucatcher/domain/expert/presentation/ExpertControllerTest.java
@@ -35,6 +35,8 @@ import com.foo.gosucatcher.domain.expert.application.dto.request.ExpertUpdateReq
 import com.foo.gosucatcher.domain.expert.application.dto.response.ExpertResponse;
 import com.foo.gosucatcher.domain.expert.application.dto.response.ExpertsResponse;
 import com.foo.gosucatcher.domain.expert.domain.Expert;
+import com.foo.gosucatcher.domain.item.application.dto.response.sub.SubItemResponse;
+import com.foo.gosucatcher.domain.item.application.dto.response.sub.SubItemsResponse;
 import com.foo.gosucatcher.domain.member.domain.Member;
 import com.foo.gosucatcher.domain.member.domain.MemberRepository;
 import com.foo.gosucatcher.global.error.ErrorCode;
@@ -303,6 +305,26 @@ class ExpertControllerTest {
 			.andExpect(status().isBadRequest())
 			.andExpect(jsonPath("$.code").value("EI001"))
 			.andExpect(jsonPath("$.message").value("해당 고수는 요청한 서비스를 등록하지 않았습니다."))
+			.andDo(print());
+	}
+
+	@Test
+	@DisplayName("해당 고수가 가진 세부 서비스 조회")
+	void getSubItemsByExpertIdSuccessTest() throws Exception {
+
+		//given
+		Long id = 1L;
+		SubItemsResponse subItemsResponse = new SubItemsResponse(
+			List.of(new SubItemResponse(1L, "알바", "청소 알바", "청소 알바 설명")));
+
+		given(expertService.getSubItemsByExpertId(id))
+			.willReturn(subItemsResponse);
+
+		//when -> then
+		mockMvc.perform(get("/api/v1/experts/{id}/sub-items", id))
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$.subItemsResponse.[0].id").value(1))
+			.andExpect(jsonPath("$.subItemsResponse.[0].name").value("청소 알바"))
 			.andDo(print());
 	}
 }
